Move _timeout.py to _utils.py
So it will be generically named, for the various generic methods that it
contains.
